refactor(uploader): stop the object browser importing the server through barrels
`src/lib/index.ts` is `export * from "./actions"` plus `export * from
"./clients"`, so importing one helper from `@/lib` pulls every server
action and every DynamoDB client behind it. Eleven client components were
doing that for a URL builder, a byte formatter, or LOGGER.

The uploader is the worst of them, and the object browser inherits it:
DirectoryList -> useUploadManager -> UploadProvider -> `@/lib`. So the
file tree on every product page carried the whole server surface into the
browser bundle to get `getTemporaryCredentials`.

Each import now names the module it actually wants -- `@/lib/logging`,
`@/lib/format`, `@/lib/urls`, `@/lib/actions/credentials`. Same for
`@/components/core`, which re-exports AccountSearchInput and reaches the
account actions through it; the five call sites here wanted MonoText,
Skeleton or CopyToClipboard.

Found while trying to give the file tree a story. That story is not here:
`@/lib/actions/credentials` is itself "use server", so DirectoryList still
reaches a server action legitimately and needs it mocked, which depends on
the mocking setup in #517. This is the half that stands on its own.

No behaviour change -- same symbols, same modules, shorter paths to them.

tsc clean · jest 65 suites / 567 tests · next lint 0 errors · storybook
build succeeds.
